move all models to models file
